Refactoring: events are first posted to the server
The game is now registering the events from the server, and sending them
to the clients.
The events are now longer directly played, but only played once they have
been validated by the server. This may induce some lag (at least one
frame) but should be OK for our purpose.