Blood effect now played in die event
New event 'die' that triggers the blood effect. This implementation is now
yet perfect since the 'die' and 'attack' event are sent in the same frame.
To fix this, the way attacks are handled should be rewritten: attack
detection should not be made in the same frame than attack action.